add broadcast.cu. broadcast_ori.cu, reducescatter.cu, reducescatter_ori.cu, modify README

+++ eval/coll/all_reduce.cu
@@ -49,7 +49,7 @@ static void getHostName(char *hostname, int maxlen) {
 }
 
 // size is the number of elements
-int run(int myRank, int nRanks, int localRank, int size, int loop,
+int run(int myRank, int nRanks, int localRank, uint64_t size, int loop,
         ncclComm_t &comm) {
   float *sendbuff, *recvbuff;
   cudaStream_t s1;
@@ -98,9 +98,9 @@ int main(int argc, char *argv[]) {
 
   assert(argc == 3);
 
-  int size = atoi(argv[1]);
+  uint64_t size = atoll(argv[1]);
   int loop = atoi(argv[2]);
-  printf("size = %d, loop = %d\n", size, loop);
+  printf("size = %lu, loop = %d\n", size, loop);
 
   int myRank, nRanks, localRank = 0;
   // initializing MPI
